validation for delete auth

+++ server/src/main/java/dataaccess/MemoryAuthDAO.java
@@ -23,6 +23,9 @@ public class MemoryAuthDAO implements AuthDAO{
 
     @Override
     public void deleteAuth(String authToken) throws DataAccessException {
+        if (authToken == null || authToken.isBlank()) {
+            throw new DataAccessException("Auth token is null or blank");
+        }
         db.remove(authToken);
     }
